move agnostic.py from utils to agnostic
